chore: rfc-index.txt minor tidyups
